Use separate testing framework package
