+++ VERSION
@@ -1,1 +1,1 @@
-2.0.0-alpha
+2.0.0-alpha-framework
