--- features/extra/no-mail.php
@@ -1,7 +1,0 @@
-<?php
-
-function wp_mail( $to ) {
-	// Log for testing purposes
-	WP_CLI::log( "WP-CLI test suite: Sent email to {$to}." );
-}
-
